+++ CHANGELOG.md
@@ -18,8 +18,14 @@ The API consists of all public Java types from `com.atlassian.performance.tools.
 ## [Unreleased]
 [Unreleased]: https://bitbucket.org/atlassian/jira-performance-tests/branches/compare/master%0Drelease-3.2.0
 
+### Added
+- Pass Jira admin credentials. Resolve [JPERF-451].
+
 ### Deprecated
 - Deprecate `RegressionResults`, because it spreads the deprecated `CohortResult`.
+- Deprecate `ProvisioningPerformanceTest` execution methods, which hardcode Jira admin credentials.
+
+[JPERF-451]: https://ecosystem.atlassian.net/browse/JPERF-451
 
 ## [3.2.0] - 2019-02-28
 [3.2.0]: https://bitbucket.org/atlassian/jira-performance-tests/branches/compare/release-3.2.0%0Drelease-3.1.0